Separate progressbar and comparing data (arrows)
